cmd/cherufe: change group names to be counter
nftables only allows 15 characters for set names, UUID is 32. Names have to be
unique only in the hypervisor so counter is ok.

The set name is implemented by having a map of FWGroupID -> N, which is later
used as a reverse lookup table to match guestIPs with the appropriate FWGroupID
set.