MIST-327 Modify and simplify the Agent interface for dover
Now that we are using loveland for placement hypervisor selection with
dover for issuing commands and updating job statuses, the Agent
interface needed to change. Pulled out the automatic status polling,
instead returning the remote job id. Removed the hypervisor candidate
checking and selection, as that's in loveland now. Expose the job status
checking method.