kappa: allow environment to override default etcd address
In order to allow environment to override defaults but not user specified
arguments and still provide the correct help message we need to do some
processing ourselves.

First initialize the variable to default or what ever is in the environment.
Then we setup pflag to return the pointer to the create variable, which we
ignore. Then we use `pflag.Visit()`, which only calls the func when an argument
has been explicitly set by the user, to save the value of the var.