Generated READMEs with godocdown
Moved package level documentation to doc.go so that it will all be
included by godoc. Generated new READMEs based on a template and
godocdown so that the same inforatmion is available easilly on github.
We now only have to maintain the in-code and doc.go documentation and
regenerate READMEs rather than manually updating in two places.

Also improved the HTTP API docs for enfield, grootslang, and waheela.