kappa: buffer etcd changes before running ansible
Changes to etcd may be bursty for config or hypervisor. Even if not too close
in time, multiple etcd changes can conceivably happen between ansible runs.
So this will buffer etcd changes and once 100ms of quiet has elapsed, ansible is
run. An upper time limit will be placed in the next commit.